Tách file code
